Add scribble_debug function for drawing text on an image for dev tests
The `scribble_debug` function has been added to draw text on an image
for checking in development. This function is used to write debug text
on an image and is useful for debugging and verifying the code during
development. The function creates a PIL ImageDraw object and uses it to
draw text on the given image.